TRAVIS Force install to be noop
Travis CI runs an install script for python, even if the install key is
not defined in `.travis.yml`. So, I need to override that invocation. In
this instance, I'm overriding it with a message about how we are
skipping it. I'll need to have travis run a CI test on this to make sure
that it is working the way I want it to.