[TRAVIS] merge all python scripts, rm docker
We have lots of python scripts all doing different things, and relying
on env vars and exit codes to track state. These changes combine all of
the python scripts into one, travis.py. The idea here is that if you
want to add functionality, add to the script.

I'm also trying to remove Docker as a dependency for testing. I did
spend a lot of time working on it, but it was a cludgy solution to a
problem that other tools can fix.